"Add GitHubPackages maven repository and new routine-client dependency"
In build.gradle, a new maven repository - GitHubPackages, has been added to handle packages stored on GitHub. This repository is associated with a particular user and repository, and requires authentication credentials for access. In addition, a new 'routine-client' project dependency has been included. These changes are necessary for accessing and implementing shared modules in the existing project using GitHub packages.